* Обновление прошедшего времени * Начало формирования отправки нового сообщения

+++ ZappChat/ZappChat/Core/Socket/Requests/AuditRequest.cs
@@ -8,7 +8,7 @@ namespace ZappChat.Core.Socket.Requests
 {
     class AuditRequest : Request
     {
-        public uint log_id { get; set; }
+        public ulong log_id { get; set; }
 
         public AuditRequest()
         {
